# CUDA: revert part of the RDNA1 optimizations