The change on the launch_bounds was causing a small performance drop in prompt processing, apparently this change was only beneficial before I tuned the mmq_y values.

| model                          |       size |     params | backend    | ngl |          test |              t/s master| t/s PR | Speedup |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: | ---------------: | ------------:
| llama 8B Q5_K - Small          |   5.21 GiB |     8.03 B | ROCm       |  99 |         pp512 |    276.60 ± 0.41 |    300.60 ± 0.46 | 1.09 |
 



- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High